requirements.txt
greenlet==3.1.1
jinja2==3.1.4
itsdangerous==2.2.0
flask==3.1.0
flask_sqlalchemy==3.1.1
blinker==1.9.0
mysql_connector_python==9.1.0
typing_extensions==4.12.2
werkzeug==3.1.3
click==8.1.7
MarkupSafe==3.0.2
SQLAlchemy==2.0.36
python_dotenv==1.0.1
pip==24.3.1
setuptools==75.6.0